Settings module created
The original django settings file has been replaced with a module which
enables different settings to be run dependent on the environment. For
example, on a local machine django will be loaded using
"--settings=settings.local".

This enables all settings to be tracked. All private information is
stored in environment variables.