Adds custom middleware
Middleware is used to enable the basket object to be passed through the request object to each view.

This checks if the user has a session variable, basket_id, then pulls
the associated basket object. If not, this is set to None so it can be
handled in the template.