Adds basket functionality
The basket enables user to add products to a virtual basket. Each
product is stored as an individual basket item which is linked to a
single basket. Items in the basket have individual quantities which
can range from 1 to 5.

The basket has an overall status stored. 'In Progress' indicates a
basket has not yet been processed, i.e. not paid for.
'Processed' indicates the basket has been paid for and will have a
correspending checkout entry for processing.

To enable easy access to a user's basket, regardless of whether they are
 logged in or not, a middleware class has been written, which adds the
user basket to the request object. This can then be accessed in views
and templates.

The basket view uses an inline formset factory to facilitate the
relationship between the Basket and BasketItem models, i.e. BasketItem
has a foreign key to Basket.

When a user logs into their account signals are used to check whether
the user added any items to their basket prior to login. If so, these
will be merged with any items in the user's stored basket, if it exists.
 If the user did not have an 'In Progress' basket, a basket is created
 and the items stored.

 Baskets are cleared upon logout, however, will be accessible upon
 logging into the user account in future.

 Robust testing added for basket views. Tests are fully documented
 within test files.