Adds BasketException and create_order method
create_order method is used when a user goes through the checkout
process. This will be called as the final step, i.e. when user payment
is successful, as indicated by Stripe. It takes all items in the user
basket and creates associated order items. Each order item is
a single item, so that items can be processed individually (the
status can be individually adjusted).

BasketException created to isolate specific issues when creating an order
from a basket.