Adds urls, views, required forms, and migrations
The checkout process captures the user's billing and shipping details.
To minimize user inputs, billing address is assumed to be the same as
those in the user profile. If not, this must be updated in the profile
first. The shipping address is also pre-populated with this information,
 however, the user is free to update this during the checkout.

 The user will be presented with a review of their order, before finally
  being asked to produce payment. An order will not be created until
  a payment has been successfully processed.

Payment processing is done via Stripe using server side and client
side processing. The server side integration setups up payment amount
and currency, and passes through a token to the front end. The token is
then used to complete the order payment when valid card details are
provided.

Once payment is verified by Stripe, the final step is run
the Basket create_order method and update order status to 'PAID'.

Base templates updated to include stripe js framework and custom
user js for payment handling.

Project urls updated to include checkout urls.