Custom user model updated
Default username and email fields updated to reflect that username is
not used, and email is used in place of a username. The Django
username_field setting changed to 'email' for normal Django (i.e. not
the allauth package). Migrations created.

This modification was required due to signup form not permitting
account creation due to a blank field being recognised as a duplicate.